+++ src/handleFailed.ts
@@ -34,10 +34,10 @@ export function handleFailed(options: Options, error: unknown, retries: number):
     // Rate limit reached
     if (error.response.status === 429) {
       const header = findHeader(error.response.headers, options.rateLimitHeaders)
-      const wait = parseResetHeader(header, Date.now(), options.resetFormat)
+      const wait = header ? parseResetHeader(header, Date.now()) : null
 
       // Add extra 1 second to account for sub second differences
-      if (wait) return wait + 1000
+      if (typeof wait === 'number') return wait + 1000
     }
 
     return backoff(retries)